azure_layer/requirements_azure.txt
azure-functions>=1.6.0
azure-cosmos>=4.0.0